[FIX] point_of_sale,pos_sale: fix product info popup
Following changes of product card that are now product.template, adapt
the product info popup to display the correct information.

Now it display information about the product template instead of the
first product variant.
