+++ addons/point_of_sale/static/tests/tours/product_configurator_tour.js
@@ -37,7 +37,7 @@ registry.category("web_tour.tours").add("ProductConfiguratorTour", {
 
             // Check that the product has been added to the order with correct attributes and price
             ProductScreen.selectedOrderlineHas(
-                "Configurable Chair (Red, Metal, Fabrics: Other: Custom Fabric)",
+                "Configurable Chair (Fabrics: Other: Custom Fabric, Metal, Red)",
                 "1.0",
                 "11.0"
             ),
@@ -50,7 +50,7 @@ registry.category("web_tour.tours").add("ProductConfiguratorTour", {
             ProductConfigurator.fillCustomAttribute("Custom Fabric"),
             Dialog.confirm(),
             ProductScreen.selectedOrderlineHas(
-                "Configurable Chair (Red, Metal, Fabrics: Other: Custom Fabric)",
+                "Configurable Chair (Fabrics: Other: Custom Fabric, Metal, Red)",
                 "2.0",
                 "22.0"
             ),
@@ -62,7 +62,7 @@ registry.category("web_tour.tours").add("ProductConfiguratorTour", {
             ProductConfigurator.pickRadio("Leather"),
             Dialog.confirm(),
             ProductScreen.selectedOrderlineHas(
-                "Configurable Chair (Blue, Metal, Leather)",
+                "Configurable Chair (Leather, Metal, Blue)",
                 "1.0",
                 "10.0"
             ),
