+++ addons/pos_sale/models/__init__.py
@@ -3,7 +3,7 @@
 
 from . import pos_config
 from . import pos_order
-from . import product_product
+from . import product_template
 from . import crm_team
 from . import pos_session
 from . import sale_order
