+++ addons/pos_sale/static/src/overrides/components/product_info_popup/product_info_popup.xml
@@ -2,7 +2,7 @@
 <templates id="template" xml:space="preserve">
     <t t-name="pos_sale.ProductInfoPopup" t-inherit="point_of_sale.ProductInfoPopup" t-inherit-mode="extension">
         <xpath expr="//div[hasclass('financials-order')]" position="before">
-            <div class="section-optional-product mt-3 mb-4 pb-4 border-bottom text-start" t-if="props.info.productInfo.optional_products.length > 0">
+            <div class="section-optional-product mt-3 mb-4 pb-4 border-bottom text-start" t-if="props.info?.productInfo?.optional_products?.length > 0">
                 <h3 class="section-title">Optional Products:</h3>
                 <div class="section-optional-product-body" t-foreach="props.info.productInfo.optional_products" t-as="optional" t-key="optional.name">
                     <span>
